made deleteUserCourses, getOrganizationCourse, and updateCourse all use the new OrgOrCourseRoles guard with tests

+++ packages/server/src/organization/organization.controller.ts
@@ -14,6 +14,7 @@ import {
   UseGuards,
   UseInterceptors,
   ParseIntPipe,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UserModel } from 'profile/user.entity';
 import { Response } from 'express';
@@ -60,7 +61,7 @@ import { ChatTokenModel } from 'chatbot/chat-token.entity';
 import { v4 } from 'uuid';
 import _ from 'lodash';
 import * as sharp from 'sharp';
-import { UserId } from 'decorators/user.decorator';
+import { User } from 'decorators/user.decorator';
 import { SchedulerRegistry } from '@nestjs/schedule';
 import { CronJob } from 'cron';
 import { OrgOrCourseRolesGuard } from 'guards/org-or-course-roles.guard';
@@ -442,18 +443,15 @@ export class OrganizationController {
   }
 
   @Patch(':oid/update_course/:cid')
-  @UseGuards(
-    JwtAuthGuard,
-    OrganizationRolesGuard,
-    OrganizationGuard,
-    EmailVerifiedGuard,
-  )
-  @Roles(OrganizationRole.ADMIN, OrganizationRole.PROFESSOR)
+  @UseGuards(JwtAuthGuard, EmailVerifiedGuard, OrgOrCourseRolesGuard)
+  @CourseRoles(Role.PROFESSOR)
+  @OrgRoles(OrganizationRole.ADMIN, OrganizationRole.PROFESSOR)
   async updateCourse(
     @Res() res: Response,
     @Param('oid', ParseIntPipe) oid: number,
     @Param('cid', ParseIntPipe) cid: number,
     @Body() courseDetails: UpdateOrganizationCourseDetailsParams,
+    @User(['organizationUser']) user: UserModel,
   ): Promise<Response<void>> {
     const courseInfo = await OrganizationCourseModel.findOne({
       where: {
@@ -583,51 +581,54 @@ export class OrganizationController {
 
     try {
       await courseInfo.course.save();
-      //Remove current profs
-      await UserCourseModel.delete({
-        courseId: cid,
-        role: Role.PROFESSOR,
-      });
-
-      for (const profId of courseDetails.profIds) {
-        const chosenProfessor = await UserModel.findOne({
-          where: { id: profId },
+      // update profs (only if admin since only admins can adjust profs in courses)
+      if (user.organizationUser.role === OrganizationRole.ADMIN) {
+        //Remove current profs
+        await UserCourseModel.delete({
+          courseId: cid,
+          role: Role.PROFESSOR,
         });
 
-        if (!chosenProfessor) {
-          return res.status(HttpStatus.NOT_FOUND).send({
-            message: ERROR_MESSAGES.profileController.userResponseNotFound,
+        for (const profId of courseDetails.profIds) {
+          const chosenProfessor = await UserModel.findOne({
+            where: { id: profId },
           });
-        }
 
-        const userCourse = await UserCourseModel.findOne({
-          where: {
-            userId: profId,
-            courseId: cid,
-          },
-        });
-
-        // user is already in the course
-        if (userCourse) {
-          userCourse.role = Role.PROFESSOR;
-          try {
-            userCourse.save();
-          } catch (err) {
-            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-              message: err,
+          if (!chosenProfessor) {
+            return res.status(HttpStatus.NOT_FOUND).send({
+              message: ERROR_MESSAGES.profileController.userResponseNotFound,
             });
           }
-        } else {
-          try {
-            await UserCourseModel.create({
+
+          const userCourse = await UserCourseModel.findOne({
+            where: {
               userId: profId,
               courseId: cid,
-              role: Role.PROFESSOR,
-            }).save();
-          } catch (err) {
-            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-              message: err,
-            });
+            },
+          });
+
+          // user is already in the course
+          if (userCourse) {
+            userCourse.role = Role.PROFESSOR;
+            try {
+              userCourse.save();
+            } catch (err) {
+              return res.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
+                message: err,
+              });
+            }
+          } else {
+            try {
+              await UserCourseModel.create({
+                userId: profId,
+                courseId: cid,
+                role: Role.PROFESSOR,
+              }).save();
+            } catch (err) {
+              return res.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
+                message: err,
+              });
+            }
           }
         }
       }
@@ -1196,18 +1197,14 @@ export class OrganizationController {
   }
 
   @Delete(':oid/drop_user_courses/:uid')
-  @UseGuards(
-    JwtAuthGuard,
-    OrganizationRolesGuard,
-    OrganizationGuard,
-    EmailVerifiedGuard,
-  )
-  @Roles(OrganizationRole.ADMIN, OrganizationRole.PROFESSOR)
+  @UseGuards(JwtAuthGuard, EmailVerifiedGuard, OrgOrCourseRolesGuard)
+  @CourseRoles(Role.PROFESSOR)
+  @OrgRoles(OrganizationRole.ADMIN, OrganizationRole.PROFESSOR)
   async deleteUserCourses(
     @Res() res: Response,
     @Param('uid', ParseIntPipe) uid: number,
     @Body() userCourses: number[],
-    @UserId() userId: number,
+    @User(['organizationUser', 'courses']) user: UserModel,
   ): Promise<Response<void>> {
     if (userCourses.length < 1) {
       return res.status(HttpStatus.BAD_REQUEST).send({
@@ -1215,23 +1212,24 @@ export class OrganizationController {
       });
     }
 
-    const userOrg = await OrganizationUserModel.findOne({
-      where: {
-        userId,
-      },
-    });
+    const isProfInAnyCourse = user.courses.some(
+      (uc) => uc.role === Role.PROFESSOR,
+    );
 
-    // If the user is just an OrganizationRole.PROFESSOR, they can only remove users from their own courses
-    if (userOrg.role === OrganizationRole.PROFESSOR) {
+    // If the user is just an org or course professor, they can only remove users from their own courses
+    if (
+      user.organizationUser.role === OrganizationRole.PROFESSOR ||
+      isProfInAnyCourse
+    ) {
       const userCoursesForUser = await UserCourseModel.find({
         where: {
-          userId: userId,
+          userId: user.id,
           courseId: In(userCourses),
         },
       });
       // all courses must be found, otherwise the user is trying to remove a course they are not in
       if (userCoursesForUser.length !== userCourses.length) {
-        return res.status(HttpStatus.UNAUTHORIZED).send({
+        return res.status(HttpStatus.FORBIDDEN).send({
           message: ERROR_MESSAGES.roleGuard.notAuthorized,
         });
       }
@@ -1242,7 +1240,7 @@ export class OrganizationController {
           userCourses.includes(courseId),
         )
       ) {
-        return res.status(HttpStatus.UNAUTHORIZED).send({
+        return res.status(HttpStatus.FORBIDDEN).send({
           message: ERROR_MESSAGES.roleGuard.notAuthorized,
         });
       }
@@ -1259,7 +1257,7 @@ export class OrganizationController {
       userInfo.role === OrganizationRole.ADMIN ||
       userInfo.organizationUser.userRole === UserRole.ADMIN
     ) {
-      return res.status(HttpStatus.UNAUTHORIZED).send({
+      return res.status(HttpStatus.FORBIDDEN).send({
         message: ERROR_MESSAGES.roleGuard.notAuthorized,
       });
     }
